feat: tcp and memory constructors for transport
